+++ gui/src/routes/+page.svelte
@@ -1,8 +1,5 @@
 <script lang="ts">
-    import {
-        DEFAULT_FILTER_FORM,
-        setFilterFormContext
-    } from '$lib/app/form-context/context'
+    import { DEFAULT_FILTER_FORM } from '$lib/app/form-context/defaults'
     import Search from '$lib/app/search/search.svelte'
 
     setFilterFormContext(DEFAULT_FILTER_FORM)
